create MarketOffer component to take createOffer method out of MarketInterface
