+++ src/app/interfaces/marketConfig.interface.ts
@@ -1,9 +1,9 @@
-import { MarketOffer } from "./marketOffer.interface";
+import { MarketOfferData } from "./marketOfferData.interface";
 
 /**
  * The list of the offers available in the Market.
  */
 export interface MarketConfig {
-  buyingOffers: Array<MarketOffer>;
-  sellingOffers: Array<MarketOffer>;
+  buyingOffers: Array<MarketOfferData>;
+  sellingOffers: Array<MarketOfferData>;
 }
